requirements.txt
PyQt5==5.15.9
pyserial==3.5
numpy==1.24.3
matplotlib==3.7.1

